Mixed parsing of new and old options
