+++ Test/lit.site.cfg
@@ -144,7 +144,7 @@ def buildCmd(args):
 dafny = addParams(buildCmd(dafnyArgs))
 # The following option adjusts the exit codes of the Dafny executable
 dafny_0 = addParams(buildCmd(['countVerificationErrors:0'] + dafnyArgs))
-standardArguments = addParams('/' + ' /'.join(dafnyArgs))
+standardArguments = addParams('--' + ' --'.join(dafnyArgs.replace(":", "="))
 
 # Inform user what executable is being used
 lit_config.note(f'Using Dafny (%dafny): {dafny}\n')
